rendering checkboxes for portfolios in ChrtGen

+++ src/WebServer/SqCoreWeb/projects/MarketDashboard/src/app/portfolio-manager/DashboardClient_portfMgr.cs
@@ -29,7 +29,7 @@ class ChartData
 {
     public ChartResolution ChartResolution { get; set; } = ChartResolution.Daily;
     public List<long> Dates { get; set; } = new List<long>();
-    public List<decimal> Values { get; set; } = new List<decimal>();
+    public List<int> Values { get; set; } = new List<int>();
 }
 
 public partial class DashboardClient
